requirements.txt
flask
requests
gunicorn
python-dotenv
mysql-connector-python